requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.3
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.1
protobuf==3.20.3
pyrsistent==0.19.3
PyYAML==6.0
requests==2.28.2
swagger-ui-bundle==0.0.9
urllib3==1.26.15
Werkzeug==2.2.3
